Adding dark/light mode persistance
When working with localStorage, it is required to take into account
server rendering because given that localStorage do not exist during
the server rendering phase it may cause problems and do not show
anything on the frontend.

In case the server rendering phase and the client rendering phase differ
then the browser will launch a warning message that needs to be
inspected. Just in case our code is well implemented we can make use of
suppressHydrationWarning property.